Create separate audit inner nav component
