+++ src/ggrc-client/js/mustache/base_objects/inner-nav.stache
@@ -3,8 +3,16 @@
     Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 }}
 
-<inner-nav
-  {instance}="instance"
-  {widget-descriptors}="innerNavDescriptors"
-  (active-changed)="showWidgetArea(%event)"
-></inner-nav>
+{{#if isAuditScope}} 
+  <audit-inner-nav
+    {instance}="instance"
+    {widget-descriptors}="innerNavDescriptors"
+    (active-changed)="showWidgetArea(%event)"
+  ></audit-inner-nav>
+{{else}}
+  <inner-nav
+    {instance}="instance"
+    {widget-descriptors}="innerNavDescriptors"
+    (active-changed)="showWidgetArea(%event)"
+  ></inner-nav>
+{{/if}}
